requirements.txt
dash==1.20.0
dash-bootstrap-components==0.10.3
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
numpy==1.18.5
pandas==1.0.5
Pillow==8.1.0
PyYAML==5.4.1
scikit-learn==0.23.1
tensorflow==2.3.0
tqdm==4.47.0